requirements.txt
Flask==2.3.0
Jinja2==3.1.2
Werkzeug==2.3.0
itsdangerous==2.1.2
MarkupSafe==2.1.2
watchdog==3.0.0
gunicorn>=21.0.0
Flask-Session==0.8.0
gspread==6.1.4
google-auth==2.36.0
openai>=1.0.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-Migrate==4.0.7
flask[async]>=2.0.0
aiohttp>=3.8.0
asgiref>=3.4.0
uvicorn>=0.23.0
flask-login>=0.6.2
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-session>=0.5.0
python-dotenv>=1.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.10.0
pandas>=2.0.0
numpy>=1.24.0
mysqlclient>=2.1.0
redis>=4.5.0